Fixed JSAPI
